+++ src/controls/MouseController.hpp
@@ -45,7 +45,7 @@ private:
     bool m_draw_cursor_for_hid_mouse = true;
     double m_last_touch_y = 0;
     
-    void handle_mouse_move(int x, int y);
+    void handle_mouse_move(double x, double y);
     void handle_mouse_buttons(int button, int action, int modifiers);
     void handle_mouse_scroll(int x, int y);
     bool set_new_mouse_state(MouseState& state);
